Extract race instance struct
